Add draft of Dict screen restructure code
- Each MDCard on Dict screen is a placeholder for set of flashcards.
- Code, connected to card list, is now separated from main part.
